Fix reference dtype handling; calibrate fp32 bias-path tolerance
The reference projected the pair embedding in the feats dtype, crashing
F.linear on mixed bf16/fp32 (and fp64-feats/fp32-weight) calls; the
embedding now casts to the weight's dtype, which is also how autocast
treats the eager module. The fp32 oracle atol splits by configuration:
sinusoid arguments reach 4096 rad where one ulp of argument difference
moves sin by ~4.9e-4 (projection-amplified to ~3e-3), so E-dependent
fp32 comparisons carry a 5e-3 floor while the no-bias path keeps 1e-6
(and passes it). Both-biases-off gradient tests now expect the
projection parameters to receive no gradient on any path.

+++ src/graphnet/models/components/flash_spacetime.py
@@ -82,7 +82,10 @@ def spacetime_pair_features(
     freqs = sinusoidal_frequencies(weight.shape[1], feats.device)
     ang = x.unsqueeze(-1) * freqs
     emb = torch.cat((torch.sin(ang), torch.cos(ang)), dim=-1)
-    return torch.nn.functional.linear(emb, weight, bias)
+    # The embedding follows the feats dtype; the projection runs in the
+    # weight's dtype, exactly as autocast treats the eager module (interval
+    # and sinusoids in fp32, Linear in the compute dtype).
+    return torch.nn.functional.linear(emb.to(weight.dtype), weight, bias)
 
 
 def pair_mask_bias(key_padding_mask: Tensor) -> Tensor:
